% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{clear_model_cache}
\alias{clear_model_cache}
\title{Clear Model Cache}
\usage{
clear_model_cache(type = "all", verbose = TRUE)
}
\arguments{
\item{type}{Character. "all", "epv", or "prediction".}

\item{verbose}{Logical. Print messages.}
}
\value{
Invisible NULL
}
\description{
Clear Model Cache
}
