% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannamodels-package.R
\docType{package}
\name{pannamodels-package}
\alias{pannamodels}
\alias{pannamodels-package}
\title{pannamodels: Pre-Trained Models for Football Analytics}
\description{
Provides pre-trained machine learning models for football analytics, served via GitHub releases with local caching. Models include xG (expected goals), xPass, EPV (expected possession value), and match prediction models. Used by the panna package for player ratings and match prediction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/peteowen1/pannamodels}
  \item Report bugs at \url{https://github.com/peteowen1/pannamodels/issues}
}

}
\author{
\strong{Maintainer}: Pete Owen \email{peteowen1@gmail.com}

}
\keyword{internal}
